# Conflicts:
#	.github/ISSUE_TEMPLATE/bug.yml
#	.github/workflows/release-update.yml
#	common-scripts.gradle
#	versions.gradle
#	vico/compose/src/main/java/com/patrykandpatrick/vico/compose/cartesian/layer/LineCartesianLayer.kt
#	vico/core/src/main/java/com/patrykandpatrick/vico/core/cartesian/layer/LineCartesianLayer.kt

+++ vico/compose/src/main/java/com/patrykandpatrick/vico/compose/cartesian/decoration/HorizontalBox.kt
@@ -26,10 +26,10 @@ import com.patrykandpatrick.vico.core.common.component.ShapeComponent
 import com.patrykandpatrick.vico.core.common.component.TextComponent
 import com.patrykandpatrick.vico.core.common.data.ExtraStore
 
-/** Creates and remembers a [HorizontalBox] */
+/** Creates and remembers a [HorizontalBox]. */
 @Composable
 public fun rememberHorizontalBox(
-  y: (ExtraStore) -> ClosedFloatingPointRange<Float>,
+  y: (ExtraStore) -> ClosedFloatingPointRange<Double>,
   box: ShapeComponent,
   labelComponent: TextComponent? = null,
   label: (ExtraStore) -> CharSequence = { HorizontalBox.getLabel(y(it)) },
